requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
mysql-connector-python==8.0.26
python-dotenv==0.19.0
Flask-WTF==1.0.0